A collection of vcache-fixes.
Added vcache to random; naive is now faster than python/numpy on all accounts!
Replaced allocation with assertion in reduce; it is no longer an extension and output is therefore already allocated by the vector engine.
Added a bh_vcache_malloc_op for use in extensions.
Removed input-check in bh_vcache_malloc(instr): The check is performed in the vector engine via assert.