bhpy: Calling bohrium.flush() now flushes.
The logic for flush(a=None) did not flush for "a=None" which is the
default when used by util.Benchmark().start() and
util.Benchmark().stop().

As a consequence, all Benchmarks have been counting the
data-initialization as part of the elapsed time.

Resulting in unfair comparison when running with/without bohrium.