npbackend: Refactoring, primarily view->View and base->Base.
The code is extremely hard to read when object-instances of "class view"
and "class base" shadows/hides/overrides the class.
Additional refactoring is done to:

 - Avoid override of built-in functions
 - Avoid variables in local scope to shadowing/hiding those in external
   scope
 - And misc. refactoring for readability according to pylint and moving
   comments of functions from # to docstring. And renaming of variables
such as "f" to "func", "v" to "view", "b" to "base".

The code-base is still very hard to read since it is quite unclear when
an argument is an "ndarray", a "View" or a "Base". When it can be one or
the other and it which cases they can be used interchangably.