core+bxx: Refined "layouts" from opcodes and BXX now uses them.
Typechecking is now entirely handled by the typechecker.
Previous several operators/operations had same-type enforced by
template parameters, these are not removed.

operator/operation generation is produced based on bytecode-layouts.
This ensures that templated functions does not exist for functions
when they do not apply.

Broadcasting is now selectively added to those marked with "broadcasting".
Strict broadcast enforcement does apply to all bytecodes, matmul could be nxmxk,
which for many cases is not broadcastable and should not be broadcast.