bxx: "Tightened" handling of implicit-result/temps and added scatterz/gatterz.
scatterz/gatterz allows for the index.len() < non-indexed-op.len().
In this case the "first" [0, ..., index.len[ of non-indexed-op is applied.

All "enqueue" operations now delete temps, this is the "tightening".
Before this change, an operation such as: bh_add(out[_(b,e)], in1, in2)
would leak the temp-view created for the output operand.