cape: Towards a nesting codegen.
The codegen currently emits patterns of "walkers", these are fixed loop constructs.
The goal is to be able to nest the output of one "walker" into the body of another
walker and thereby facilitate "nesting" of iteration spaces and generic "streaming"
in contrast to special-case "streaming".

Added Codeblock and Loop emitters.